+++ lib/src/features/assessments/domain/models/assessment/content/paper_type.dart
